requirements.txt
streamlit
google-generativeai
python-dotenv #load environment variables
langchain
PyPDF2 #load any pdf or to read any pdf
chromadb #specifically for vector embeddings
